move StaticChoiceProvider outside of test module
